+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.12.16
+
+* `Coroutines`:
+  * `Android`:
+    * Add class `FlowOnHierarchyChangeListener`
+    * Add `ViewGroup#setOnHierarchyChangeListenerRecursively(OnHierarchyChangeListener)`
+
 ## 0.12.15
 
 * `Common`: